fix(dx): harden install, package cbh as pipx-installable, fix broken commands
Addresses 6 P1 defects from a DX review (solo bug-bounty hunter persona):

- install: .gitattributes (eol=lf) + CRLF self-heal in install.sh; INSTALL uses bash form
- cli: package cbh as a real pip/pipx console-script (pyproject.toml + cbh/ package +
  bundled cbh/data/skill_index.json fallback + scripts/gen_skill_index.py); scripts/cbh.py
  kept as a clone shim; recon writes to ./recon when installed
- commands: add /scope backed by a new deterministic engine/scope.py CLI (was called
  mandatory in /hunt but missing); reframe /token-scan, /intel, /surface so the skill
  path is primary and missing helper scripts are clearly optional
- docs: install capability matrix (plugin vs clone vs pipx); reconcile the authorization
  story in /hunt (don't ingest the SOW document != skip scope verification)
