+++ cbh/__init__.py
@@ -1,0 +1,2 @@
+"""claude-bughunter CLI package."""
+__version__ = "2.1.0"
