+++ commands/intel.md
@@ -1,6 +1,6 @@
 ---
 name: intel
-description: On-demand intelligence fetch for a target — CVEs, disclosed reports, new features. Wraps learn.py + hunt memory context. Usage: /intel target.com
+description: On-demand intelligence fetch for a target — CVEs, disclosed reports, new features. Pulls NVD/GitHub-Advisory CVEs + bundled disclosed reports + hunt memory context. Usage: /intel target.com
 ---
 
 # /intel
@@ -9,8 +9,10 @@ Fetch actionable intelligence for a target.
 
 ## What This Does
 
-1. Runs `learn.py` for CVEs and advisories matching the target's tech stack
+1. Looks up CVEs and advisories matching the target's tech stack (NVD + GitHub
+   Advisory via web search; `learn.py` is used as an accelerator only if present)
 2. Fetches HackerOne Hacktivity for the target (via HackerOne MCP if available)
+   and cross-references the bundled `docs/disclosed-reports/`
 3. Cross-references with hunt memory — flags untested CVEs and new endpoints
 4. Outputs prioritized intel with hunt recommendations
 
@@ -48,8 +50,11 @@ MEMORY CONTEXT:
 
 | Source | What | Auth required? |
 |---|---|---|
-| `learn.py` — NVD | CVEs matching tech stack | No |
-| `learn.py` — GitHub Advisory | Security advisories | No |
-| `learn.py` — HackerOne Hacktivity | Disclosed reports | No |
-| HackerOne MCP (if connected) | Program stats, policy | No (public) |
+| NVD (web) | CVEs matching tech stack | No |
+| GitHub Advisory (web) | Security advisories | No |
+| `docs/disclosed-reports/` | Bundled disclosed-report patterns | No (local) |
+| HackerOne Hacktivity / MCP (if connected) | Disclosed reports, program stats | No (public) |
 | Hunt memory | Previously tested endpoints | Local files |
+
+> `learn.py` is an optional accelerator. If it isn't present, Claude performs the
+> lookups above directly — the command works without it.
